+++ module/runtime/test/binary.ts
@@ -1,3 +1,4 @@
+import crypto from 'node:crypto';
 import assert from 'node:assert';
 import { Readable } from 'node:stream';
 import { buffer } from 'node:stream/consumers';
@@ -53,5 +54,23 @@ export class BytesUtilTest {
     assert(blob.size === undefined);
     assert(blobBytes.length === allBytes.length);
     assert(blobBytes.equals(allBytes));
+  }
+
+
+  @Test()
+  async verifySimpleHash() {
+    const hash = crypto.createHash('sha512');
+    hash.update('roger');
+    const key = hash.digest('hex');
+
+    assert(BinaryUtil.hash('roger', 64) === key.substring(0, 64));
+
+    const hash2 = crypto.createHash('sha512');
+    hash2.update('');
+    const unKey = hash2.digest('hex');
+
+    assert(BinaryUtil.hash('', 20) === unKey.substring(0, 20));
+
+    assert(BinaryUtil.hash('', 20) !== key.substring(0, 20));
   }
 }
