+++ module/runtime/test/util.ts
@@ -1,4 +1,3 @@
-import crypto from 'node:crypto';
 import assert from 'node:assert';
 
 import { Test, Suite } from '@travetto/test';
@@ -16,23 +15,6 @@ export class UtilTest {
 
     assert(Util.uuid().length === 32);
     assert(/^[0-9a-f]{32}$/.test(Util.uuid(32)));
-  }
-
-  @Test()
-  staticUuidVerify() {
-    const hash = crypto.createHash('sha512');
-    hash.update('roger');
-    const key = hash.digest('hex');
-
-    assert(Util.hash('roger', 64) === key.substring(0, 64));
-
-    const hash2 = crypto.createHash('sha512');
-    hash2.update('');
-    const unKey = hash2.digest('hex');
-
-    assert(Util.hash('', 20) === unKey.substring(0, 20));
-
-    assert(Util.hash('', 20) !== key.substring(0, 20));
   }
 
   @Test()
